+++ app/src/components/forms/__tests__/DepositBankForm.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils'
 import { describe, it, expect } from 'vitest'
-import DepositBankForm from '@/components/modals/DepositBankForm.vue'
+import DepositBankForm from '@/components/forms/DepositBankForm.vue'
 import { NETWORK } from '@/constant'
 
 describe('DepositBankModal.vue', () => {
